+++ src/NodeCollection.php
@@ -34,7 +34,7 @@ class NodeCollection implements \ArrayAccess, \Countable, \IteratorAggregate
         return (new NodeCollectionFactory())->fromIterable(
             new static(),
             $entries,
-            $createNode ?? fn ($entry) => $entry instanceof Node ? $entry : new DefaultNode($entry)
+            $createNode ?? fn ($entry) => $entry instanceof Node ? $entry : new DefaultNode($entry, new static())
         );
     }