+++ tests/Fixtures/CustomNodeCollection.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Thinktomorrow\Vine\Tests\Fixtures;
+
+use Thinktomorrow\Vine\NodeCollection;
+
+class CustomNodeCollection extends NodeCollection
+{
+
+}